Experimental jsi18n/ caching
Adapted from:
https://docs.djangoproject.com/en/5.0/topics/i18n/translation/#note-on-performance
NOTE: When testing locally, this suspiciously tends to perform worse for
some reason (~14ms for normal response, ~56ms for 304). But on an
instance behind NGINX this performs better on average (~150ms for
normal, ~40ms for 304).
As to why it prefers to GET for 304 istead of caching, I have no clue.
My guess is the Vary header. That's a TODO.
Probably works better in normal deployment with NGINX.
We probably should consider django-statici18n like the documentation
suggests, though that project seems to be a bit dormant.